rename sql unit test
