+++ .gitignore
@@ -10,3 +10,4 @@
 /logs
 sql/my_dnn_model
 .DS_Store
+.vscode
